Add linux-next merge script and list of trees
Add a simple script to iterate over trees that are to be included in the
linux-next tree and merge them. Note that the initial branch needs to be
set up manually (latest -rc from Linus should be good). When a conflict
is detected, the script will drop you to a subshell to let you resolve
the conflict. To resume the merge, exit the subshell.

Note that this script doesn't handle the mmotm "tree" yet.

The list of trees is taken from next-20130927 (with the URL of the VFS
tree fixed up).

Signed-off-by: Thierry Reding <treding@nvidia.com>