Add git-topic script
This script is useful for managing a set of topic branches. It works by
keeping track of a common base for topic branches and can rebase all of
them on top of a new common base.

The command uses the staging/ prefix for branches and it is recommended
that topic branches are created in the same namespace (though they don't
necessarily have to be). Three branches are special: base, master and
work. base is used to track the common base of all topic branches. After
rebasing onto a new base the command will merge topic branches one by
one into the master branch. The work branch is based on master and will
typically contain work-in-progress commits that haven't been pulled into
separate topic branches yet.

The list of topic branches is maintained in .git/topic-branches.

Signed-off-by: Thierry Reding <treding@nvidia.com>