git-topic: Add push command, dry-run support and fixes
This fixes a number of minor issues with the rebase command, such as
properly removing intermediate branches and tags to automatically keep
track of the last rebase. In addition, the rebase command now supports
dry-run mode which prints what it would do without changing the index.

A new push command is added to push all topic branches to a remote
repository.

Signed-off-by: Thierry Reding <treding@nvidia.com>