git-topic: Rebase onto base by default
If no target branch was given to the rebase command, rebase onto the
base branch by default. This won't actually rebase the individual topic
branches because the base hasn't changed, but it will cause the branches
to be merged into the master branch and rebase the work branch on top of
that.

Signed-off-by: Thierry Reding <treding@nvidia.com>