drm: Create build.log and error.log, support colors
Write the build output to build.log in the output directory and the
output of stderr to error.log. The error.log file will be used to check
if any warnings were encountered and report status differently.

Also add colorized output support (defaults to off, can be enabled with
the -c, --color option) to make it visually more obvious what the build
status is.

Signed-off-by: Thierry Reding <treding@nvidia.com>