tegra-kernel: Comment out Tegra210-specific installation
Tegra210 doesn't currently support network boot out of the box because
the Ethernet device is on the USB bus that U-Boot doesn't support yet.
The typical installation procedure would therefore involve running the
UMS command within U-Boot to expose a mass storage device to the build
system. However, that's very system-specific and not very reliable.

Instead, the recommended approach is to use a PCI Ethernet device to
work around the current limitation so that no custom installation is
necessary.

Signed-off-by: Thierry Reding <treding@nvidia.com>